requirements.txt
Flask==2.3.2
easyocr
torch
pillow
numpy